fix ExpressionChangedAfterItHasBeenCheckedError
I think we should call `updateFilterWidth` in `ngOnInit` instead of `ngAfterViewInit`. 

The original value (1) of `filterInputWidth` gets bound to the underlying input element before `ngAfterViewInit` is called. If we call  `updateFilterWidth`, which updates `filterInputWidth`, in `ngAfterViewInit` we will change a bound input property in the middle of a change detection loop, which angular should detect in dev mode and throw ExpressionChangedAfterItHasBeenCheckedError.