Filter box in multiple select
It's a simple modification that seems to give back the filter box to multiple select. I didn't investigate further in the history of changes you made, so maybe there is a reason to remove filter from multiple select that I don't understand. Anyway, thank you for this great work, till now this is the best plugin I've found for select.